pyproject.toml
[tool.poetry]
name = "divvy-analysis"
version = "0.1.0"
description = ""
authors = ["Joe Gambino <joe@gam.bingo>"]
readme = "README.md"
packages = [{include = "divvy_analysis"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
streamlit = "^1.31.0"
routingpy = "^1.3.0"
osmnx = "^1.9.1"
matplotlib = "^3.8.2"
scikit-learn = "^1.4.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.27"
selenium = "^4.18.1"
tqdm = "^4.66.2"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.0.12"
ipykernel = "^6.29.0"
watchdog = "^3.0.0"
cartopy = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
